Struct samp::prelude::Ref

pub struct Ref<'amx, T> where
    T: AmxPrimitive,  { /* fields omitted */ }

A reference to a cell in the Amx.

Notes

This type implements Deref trait that allows you read or write to inner value (like smart pointers Box<T>, Rc<T>).

Methods

impl<'amx, T> Ref<'amx, T> where
T: AmxPrimitive,

pub unsafe fn new(amx_addr: i32, phys_addr: *mut T) -> Ref<'amx, T>

Create a new wrapper over an AMX cell.

Safety

Ref<T> should be alive as long as phys_addr or it will dangling pointer.

It's not recomended to use directly, instead get a reference from Args or Amx::get_ref.

pub fn address(&self) -> i32

Get an inner AMX address to cell (not physical).

Example

use samp_sdk::cell::Ref;
fn native_fn(amx: &Amx, arg: Ref<usize>) {
    let cell_addr = arg.address();
    println!("The argument stored in the {} cell.", cell_addr);
}

pub fn as_ptr(&self) -> *const T

Get a pointer to a memory cell.

pub fn as_mut_ptr(&mut self) -> *mut T

Get a mutable pointer to a memory cell.